requirements.txt
# Core
pytz==2023.3
Django==4.2.1
django-configurations==2.4.1
gunicorn==20.1.0
newrelic==8.8.0
python-decouple==3.8
daphne==4.0.0
channels==4.0.0

# For the persistence stores
psycopg2-binary==2.9.6
dj-database-url==2.0.0

# For the cache
django-redis==5.4.0

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
djoser==2.2.2
django-filter==23.2
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.1
drf-base64==2.0
djoser==2.2.2

# Documentation
drf-spectacular==0.27.0

# Developer Tools
ipdb==0.13.13
mkdocs==1.4.3
flake8==6.0.0
django-jazzmin==2.6.0
black==23.12.1
isort==5.13.2
pre-commit==3.6.0

# Testing
mock==5.0.2
factory-boy==3.2.1
django-nose==1.4.7
coverage==7.2.5
django-debug-toolbar==4.2.0

# Static and Media Storage
django-storages==1.13.2
boto3==1.26.135
Pillow==10.0.0
whitenoise==6.5.0
